Added more complex NPC interactions, building framework for quests.
