+++ Sprites.py
@@ -311,122 +311,6 @@ class Mob(pg.sprite.Sprite):
         self.rect.center = self.hit_rect.center
 
 
-class NonPlayerCharacter(pg.sprite.Sprite):
-    def __init__(self, game, x, y):
-        self.groups = game.all_sprites, game.npcs
-        pg.sprite.Sprite.__init__(self, self.groups)
-        self.game = game
-        self.images = {
-            "NPC_r": self.game.spritesheet_k_r.get_image(0, 0, 100, 100)
-        }
-        for image in self.images:
-            self.images[image].set_colorkey(BG_SPRITE_COLOR)
-        self.image = self.images["NPC_r"]
-        self.rect = self.image.get_rect()
-        self.rect.center = (x, y)
-        self.hit_rect = self.rect.copy()
-        self.hit_rect.center = self.rect.center
-        self.pos = vec(x, y)
-        self.vel = vec(0, 0)
-        self.acc = vec(0, 0)
-        self.rect.center = self.pos
-        self.rot = 0
-        self.health = NPC_HEALTH
-        self.click_delay = pg.time.get_ticks()
-        self.active = False
-
-        # Specific variables for this NPC
-        self.health_packs = 1
-
-    def get_clicked(self):
-        mouse = pg.mouse.get_pressed()
-
-        # Check distance from player
-        now = pg.time.get_ticks()
-        dist = self.pos - self.game.player.hit_rect.center
-        if mouse[0]:
-            if self.game.camera.apply_rect(self.hit_rect).collidepoint(pg.mouse.get_pos()):
-                if now - self.click_delay > 1000:
-                    self.click_delay = now
-                    if 0 < dist.length() < SPEAK_RANGE:
-                        self.active = True
-        if dist.length() > SPEAK_RANGE:
-            self.active = False
-            self.reset_dialog()
-
-    def reset_dialog(self):
-        self.game.dialog = False
-        self.game.dialog_selection = None
-        self.game.dialog_text = ""
-        self.game.dialog_options = []
-        self.active = False
-
-    def dialog_text(self):
-        text = ""
-        options = []
-        if self.health_packs > 0:
-            text = "It is dangerous to go alone, take this"
-            options = ["Thanks", "Nod"]
-        else:
-            text = "You have already taken my health pack, back off man"
-            options = ["..."]
-
-        return text, options
-
-    def dialog_0(self):
-        if self.health_packs > 0:
-            self.health_packs -= 1
-            self.game.player.add_item("Health")
-
-    def dialog_1(self):
-        if self.health_packs > 0:
-            self.health_packs -= 1
-            self.game.player.add_item("Health")
-
-    def dialog_2(self):
-        pass
-
-    def dialog_3(self):
-        pass
-
-    def dialog_4(self):
-        pass
-
-    def update(self):
-        self.get_clicked()
-        if self.health <= 0:
-            self.kill()
-        # self.rot = (self.game.player.pos - self.pos).angle_to(vec(1, 0))
-        self.rect = self.image.get_rect()
-        self.rect.center = self.pos
-        self.acc = vec(NPC_SPEED, 0).rotate(-self.rot)
-        self.acc += self.vel * -1
-        self.vel += self.acc * self.game.dt
-        self.pos += self.vel * self.game.dt + 0.5 * self.acc * self.game.dt ** 2
-        self.hit_rect.centerx = self.pos.x
-        if self.vel.x >= 0:
-            self.image = self.images["NPC_r"]
-        else:
-            self.image = self.images["NPC_r"]
-        collide_with_walls(self, self.game.walls, 'x')
-        self.hit_rect.centery = self.pos.y
-        collide_with_walls(self, self.game.walls, 'y')
-        self.rect.center = self.hit_rect.center
-
-        if self.active:
-            if self.game.dialog_selection is None:
-                self.game.dialog = True
-                self.game.dialog_text, self.game.dialog_options = self.dialog_text()
-            elif self.game.dialog_selection == 0:
-                self.dialog_0()
-                self.reset_dialog()
-            elif self.game.dialog_selection == 1:
-                self.dialog_1()
-                self.reset_dialog()
-            else:
-                self.reset_dialog()
-
-
 class Obstacle(pg.sprite.Sprite):
     def __init__(self, game, x, y, w, h):
         self.groups = game.walls
